integrate standart and flat price
